requirements.txt
python-telegram-bot==13.15
deep-translator
python-docx
PyPDF2